### 2023-03-14 379258c
update missing doi per JOSS reference check
