+++ paper.bib
@@ -31,7 +31,7 @@ volume = {104},
 pages = {157-162},
 year = {2019},
 issn = {0308-597X},
-doi = {https://doi.org/10.1016/j.marpol.2019.02.019},
+doi = {10.1016/j.marpol.2019.02.019},
 url = {https://www.sciencedirect.com/science/article/pii/S0308597X18306675},
 author = {Kimberley T.A. Davies and Sean W. Brillant}
 }
@@ -60,7 +60,8 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   volume={135},
   number={1},
   pages={1--31},
-  year={2019}
+  year={2019},
+  doi={10.3354/dao03376}
 }
 
 @article{CookeIUCN2020,
@@ -83,7 +84,8 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   number={9},
   pages={1476--1489},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/2041-210x.13244}
 }
 
 @article{parks2011sound,
@@ -93,7 +95,8 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   volume={15},
   number={1},
   pages={63--76},
-  year={2011}
+  year={2011},
+  doi={10.3354/esr00368}
 }
 
 @article{corkeron2018recovery,
@@ -123,7 +126,8 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
 	  title = {Ocean Regime Shift is Driving Collapse of the North Atlantic Right Whale Population},
 	  journal = {Oceanography},
 	  year = {2021},
-	  month = {September}
+	  month = {September},
+	  doi={10.5670/oceanog.2021.308}
 	}
 
 @article{johnson2021whalemap,
@@ -133,7 +137,8 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   volume={6},
   number={62},
   pages={3094},
-  year={2021}
+  year={2021},
+  doi={10.21105/joss.03094}
 }
 
 @electronic{NMFS:2008,
@@ -166,5 +171,6 @@ Atlantic right whales}",
   volume={8},
   pages={760840},
   year={2021},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi={10.3389/fmars.2021.760840}
 }
