### 2023-03-23 #5254 @kthyng
Hi @leahcrowe! I think you just need to have the command at the top of a message.
