### 2023-05-03 #5436 @editorialbot
[REVIEW]: Shiny tools for management rules: interactive applications that aid in conservation   strategies for North Atlantic right whales
**Submitting author:** @leahcrowe (<a href="http://orcid.org/0000-0001-9133-8522">Leah Crowe</a>)
**Repository:** https://github.com/NEFSC/READ-PSB-LWT-narwss_rwsas_apps
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
