### 2023-05-07 dd4417d
added missing DOIs
