+++ paper.bib
@@ -21,7 +21,8 @@
   journal={Endangered Species Research},
   volume={45},
   pages={251--268},
-  year={2021}
+  year={2021},
+  doi={10.3354/esr01137}
 }
 
 @article{DAVIES2019157,
@@ -41,7 +42,8 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   title         = "{North Atlantic Right Whale Consortium 2021 Annual Report Card}",
   year          = "2022",
   howpublished = "Report to the North Atlantic Right Whale Consortium",
-  url           = "https://www.narwc.org/uploads/1/1/6/6/116623219/2021report_cardfinal.pdf"
+  url           = "https://www.narwc.org/uploads/1/1/6/6/116623219/2021report_cardfinal.pdf",
+  doi = {10.1575/1912/66099}
 }
 
 @electronic{vesselspeedrule2020,
@@ -107,7 +109,8 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   number={11},
   pages={180892},
   year={2018},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
+  doi={10.1098/rsos.180892}
 }
 
 @article{davis2017long,
@@ -118,7 +121,8 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   number={1},
   pages={1--12},
   year={2017},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41598-017-13359-3}
 }
 
 @article{Meyer-Gutbrod:2021,
