### 2023-05-09 #5436 @TanyaS08
I'm going to attach/insert myself here as well since I had the same issue (I also added some dummy digits for a flight number - same message). I was wondering if it might also be an OS issue when setting the path? I'm running macOS for what it's worth...
