### 2023-05-09 #5436 @leahcrowe
Hi all,

Thank you for agreeing to test this app. We are grateful for your involvement and review!

You don't need to include a tail number to proceed, but I have updated the wiki page with an example tail number: https://github.com/NEFSC/READ-PSB-LWT-narwss_rwsas_apps/wiki/Aerial-Survey-Processing-App,-Aerial-Survey-Tab:-Part-1

The issue you all were running into was that I incorrectly named the folder containing the example data (was YYYYMMDD when it should have been YYMMDD). All good now, and I have updated the README file with how to run the example data (without needing to download anything to your machine).

I'm aware that you all will not be able to test the trigger evaluation without access to our database server, so I am working on an example workaround.

Hopefully at least the processing app goes a bit more smoothly for you -- sorry about that mix up!

Leah
