### 2023-05-09 #5436 @salix-d
The problem was that I thought the path to the files included the folder `210409` and/or that they needed to be in a folder named with the date.

Perhaps " the pathway you enter should be the pathway where the edit files are stored." could be change to something like " the pathway you enter should be the pathway to the folder containing the edit files are stored. This file should also be labeled like [YYMMDD]."
