### 2023-05-09 #5436 @salix-d
For the dependencies, I notice that if I don't load the libraries, they get loaded when doing `shiny::runGitHub("READ-PSB-LWT-narwss_rwsas_apps", username = "NEFSC", ref = "master")` so the user doesn't have to copy the list from the readme.

Since the review check list mentionned that
> Ideally these should be handled with an automated package management solution.

I would suggest using `require` instead of `library` in the [global_libraries.R](https://github.com/NEFSC/READ-PSB-LWT-narwss_rwsas_apps/blob/master/scripts/global_libraries.R) script. This will install any missing packages for the user.
