+++ scripts/global_libraries.R
@@ -1,37 +1,37 @@
 ##load libraries
-library(shiny)
-library(shinyjs)
-library(shinydashboard)
-library(dplyr)
-library(ggplot2)
-library(rgeos)
-library(rgdal)
-library(sp)
-library(raster)
-library(scales)
-library(extrafont)
-library(fontcm)
-library(leaflet)
-library(leaflet.esri)
-library(knitr)
-library(lubridate)
-library(data.table)
-library(zoo)
-library(rhandsontable)
-library(RColorBrewer)
-library(reshape)
-library(geosphere)
-library(maptools)
-library(maps)
-library(rmarkdown)
-library(tinytex)
-library(igraph)
-library(htmlwidgets)
-library(webshot)
-library(rlist)
-library(ellipsis)
-library(stringi)
-library(RODBC)
-library(DBI)
-library(sqldf)
-library(shinymanager)
+require(shiny)
+require(shinyjs)
+require(shinydashboard)
+require(dplyr)
+require(ggplot2)
+require(rgeos)
+require(rgdal)
+require(sp)
+require(raster)
+require(scales)
+require(extrafont)
+require(fontcm)
+require(leaflet)
+require(leaflet.esri)
+require(knitr)
+require(lubridate)
+require(data.table)
+require(zoo)
+require(rhandsontable)
+require(RColorBrewer)
+require(reshape)
+require(geosphere)
+require(maptools)
+require(maps)
+require(rmarkdown)
+require(tinytex)
+require(igraph)
+require(htmlwidgets)
+require(webshot)
+require(rlist)
+require(ellipsis)
+require(stringi)
+require(RODBC)
+require(DBI)
+require(sqldf)
+require(shinymanager)
