+++ README.md
@@ -35,7 +35,7 @@ To take a spin with processing example aerial survey data and evaluating if sigh
 
 210409: Visual sightings that extend two active and overlapping SLOW Zones. Two flight day, no editted eff/sig file.
 
-210512: Visual sightings that do not trigger new protection measures. One flight day, includes option to load previously editted eff/sig file.
+210512: Visual sightings that fall within active SLOW Zones, but trigger no further action. One flight day, includes option to load previously editted eff/sig file.
 
 ## Script Flow Chart
 
