### 2023-05-10 #5436 @leahcrowe
Hi all,

I've added in three more example days of data that offer different scenarios for the trigger analysis from the aerial survey data (see the README file for the update). For our functional operations, we have this app interacting on our internal server and database, so I can't offer up that actual experience, but I hope the examples I have provided demonstrate how this app is going through the decision making process for triggering protection zones, or not. Please let me know if there is anything further you would like to see to proceed.

Leah
