+++ paper.md
@@ -46,7 +46,7 @@ The vessel speed restriction ruling was first put into place in late 2008 [@NMFS
 
 The expansion of the dynamic management conservation strategy, as well as the increase in detection of right whales outside of SMAs, called for tools to: 1) streamline the process of handling multiple streams of right whale detection data, 2) refine and modernize the process of determining if the trigger criteria for dynamic measures were met, 3) quickly and accurately calculate the boundaries of a proposed Slow Zone, and 4) communicate this information to federal managers in near real-time.
 
-The North Atlantic Right Whale Sighting Survey (NARWSS) 'Aerial Survey Data Processing' tool, and the Right Whale Sighting Advisory System (RWSAS) 'Trigger Analysis' tool improve the workflow between data processing and the determination of potential Right Whale Slow Zones (Fig. 2). These tools provide a platform to efficiently process survey data to help eliminate sources of human error, integrate currently active management zones, determine if trigger criterion for a new dynamic protection zone have been met, calculate the boundaries of a proposed Slow Zones, and compile reports that are then sent on to managers and other stakeholders. 
+The North Atlantic Right Whale Sighting Survey (NARWSS) 'Aerial Survey Data Processing' tool, and the Right Whale Sighting Advisory System (RWSAS) 'Trigger Analysis' tool improve the workflow between data processing and the determination of potential Right Whale Slow Zones (Fig. 2). These tools provide a platform to efficiently process survey data to help eliminate sources of human error, integrate currently active management zones, determine if trigger criterion for a new dynamic protection zone have been met, calculate the boundaries of a proposed Slow Zone, and compile reports that are then sent on to managers and other stakeholders. 
 
 The tools described here were developed for scientists at the Northeast Fisheries Science Center (NEFSC) in their role as data stewards to support federal managers at the Greater Atlantic Regional Fisheries Office; however, the data processing and reporting procedure will be of interest to wildlife survey and management teams with similar objectives. Additionally, these publicly available tools provide transparency and reproducibility of the process for declaring dynamic protection zones in the Northeast United States. These tools were written primarily using the R coding language [@Rlanguage] and specifically leverage the 'Shiny' package [@shiny].
 
@@ -56,7 +56,7 @@ Current aerial survey data collection software used by the NEFSC renders three r
 
 # 'Trigger Analysis'
 
-The 'Trigger Analysis' tool evaluates whether a dynamic Slow Zone should be declared based on either visual or acoustic detections. The tool accesses detections from several sources, including governmental, educational, military, commercial, and non-profit institutions, in addition to the general public, which are stored in internal and external databases. Visual and acoustic detection data are evaluated on a daily scale to determine if the trigger criterion has been met. Acoustic slow zones are triggered when three or more upcalls (an acoustic vocalization that is common for this species and is known to be made by all ages and sex classes [e.g. @parks2011sound; @van2021noaa]) are acoustically detected within a 15-minute period, and DMAs are trigger when a minimum density of 4 whales per 100 nm$^2$ are observed. Both criterion must occur spatially outside of active SMAs and dynamic Slow Zones to trigger a new area, and can potentially extend the time period of active dynamic zones (see Fig. 3 for the decision making process built into the tool). When a dynamic Slow Zone is triggered, this tool will determine the boundaries of the zone and name it relative to the closest port or landmark. Both a letter to federal managers proposing the zone, as well as a report describing the potential zone, are generated as PDFs, and the data describing the zone are uploaded to an internal database. 
[…]
