### 2023-05-10 #5436 @TanyaS08
> > I was able to download the report in html but not in pdf format.
> > I had this error :
> > > An error has occurred!
> > > LaTeX failed to compile C:\Users\salix\AppData\Local\Temp\RtmpwVL36t\file39543fa5181d.tex. See https://yihui.org/tinytex/r/#debugging for debugging tips. See file39543fa5181d.log for more info.
> >
> >
> > Here's the log file: [file39543fa5181d.log](https://github.com/openjournals/joss-reviews/files/11436903/file39543fa5181d.log)
> > Don't know if related, but the html file was named "09Apr2021_NOAA_NERW_Aerial_Report.html" and the pdf file "report.htm".
>
> This is an issue that occurs related to tinytex/your machine's LaTeX distribution. I offer up some tests/installation instructions in the "Additional installation" README section. I have been able to set this up on several different computers and our shiny server, but some computers have eluded me, which is why we have offered the html download option. The correct naming convention is added to the file when it works properly.

I've also had an issue with creating the PDF (but everything else works! 😊). My error message looks a bit different though and says the following:
I was unable to find any missing LaTeX packages from the error log file947ed0e2974.log.
! Undefined control sequence.
l.81 title: ``NOAA Northeast Region \nRight
                                            Whale Aerial Survey Report''

which is then followed by:

Warning: Error in : LaTeX failed to compile /var/folders/wf/wmp7x6xj7_zbx2bl9z5ysfwm0000gn/T//Rtmpuirwmm/file947ed0e2974.tex. See https://yihui.org/tinytex/r/#debugging for debugging tips. See file947ed0e2974.log for more info.

It could be that I also have an underlying LATEX issue but I was wondering about the `l.81 title: ``NOAA Northeast Region \nRight  Whale Aerial Survey Report'' ` comment...
