+++ scripts/NARWSSglobal.R
@@ -1,40 +1,0 @@
-##load libraries
-
-library(shiny)
-library(shinyjs)
-require(dplyr)
-library(ggplot2)
-library(rgeos)
-library(rgdal)
-library(sp)
-library(raster)
-library(scales)
-library(extrafont)
-library(fontcm)
-library(leaflet)
-require(knitr)
-require(lubridate)
-require(data.table)
-require(zoo)
-library(leaflet.esri)
-library(rhandsontable)
-library(RColorBrewer)
-library(reshape)
-library(geosphere)
-library(maptools)
-library(maps)
-library(measurements)
-library(rmarkdown)
-library(RODBC)
-library(DBI)
-library(sqldf)
-library(tinytex)
-library(igraph)
-library(htmlwidgets)
-library(webshot)
-
-#devtools::install_github("jrowen/rhandsontable")
-#devtools::install_github('yihui/tinytex')
-#devtools::install_github('rstudio/rmarkdown')
-#devtools::install_version('rmarkdown', version = '1.8')
-#webshot::install_phantomjs()
