### 2023-05-11 NEFSC/READ-PSB-LWT-narwss_rwsas_apps#26 @TanyaS08
Documenting Functionality
This is primarily relating to the following point in the reviewer checklist regarding Documentation

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

To me I feel that a core argument justification of releasing this software is the idea that it can by modified and used for other monitoring applications/setting (I love the fact that this type of app development is being shared for re-use 🐣). However, since this means re-users/potential modifiers might be quite daunted by where to start in terms of what script does what I wonder if it could be a bit more extensively documented? I think script flow chart figure is a great starting point but it might be nice to also attach some sense of functionality as to what each script/group of scripts does. This could be something simple like scripts x,y,z deal with the raw sightings data, I,j,k, with report generation etc. Do the colours in the figure maybe represent that already?

I don't want to generate undue work though so I'd be interested to see what @salix-d thoughts are and if they think that the documentation is sufficient (or have other thoughts). Of course as the authors y'all should also have a say!
