### 2023-05-11 NEFSC/READ-PSB-LWT-narwss_rwsas_apps#26 @salix-d
The arrows/their colors do seem toindicate which app is being used, and each app seems to have its own functionnality.

I also like that the code is commented. It could be beneficial to have roxygen type of documentation in the files, like defining the arguments of each function. That does help contributors/re-users, but I don't know if it's _necessary_. I don't see many packages that have documentation for _internal_ functions.

Kind of related to that though, I don't think I saw community guidelines
> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
