### 2023-05-12 NEFSC/READ-PSB-LWT-narwss_rwsas_apps#26 @TanyaS08
Okay I can see an argument that the figure does do enough lifting on its own. However may I then suggest that some consideration be given to the colours of the arrows as the colours themselves are not very divergent from a colour vision deficiency perspective. Perhaps an an additive to the colours different arrowheads (as you did for the figure in the actual manuscript) could be used.
