### 2023-05-12 NEFSC/READ-PSB-LWT-narwss_rwsas_apps#26 @salix-d
Using different line types is also a great option.
I would suggest a legend as well and having a pdf/text version for screen readers.
