### 2023-05-17 NEFSC/READ-PSB-LWT-narwss_rwsas_apps#26 @leahcrowe
> Kind of related to that though, I don't think I saw community guidelines
>
> > Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Thanks for catching the absence of community guidelines. I added some more details regarding third party contributions, issues, and support at the beginning of the README
