### 2023-05-17 fdbb9ff
Update README.md
define script structure
