### 2023-05-17 NEFSC/READ-PSB-LWT-narwss_rwsas_apps#26 @leahcrowe
Thank you both for your comments and considerations regarding the description of the script structure and specifically the flow chart figure. I considered your comments and decided to include both a short description of each file as well as an updated version of the flow chart. Let me know if you have any suggestions after considering these changes!
