### 2023-05-23 NEFSC/READ-PSB-LWT-narwss_rwsas_apps#26 @salix-d
> I added some more details regarding third party contributions, issues, and support at the beginning of the README

Hi, I did see the line you added. I think it could be more precise for the third party contribution. Often when contributions are accepted there's a code of conduct or some kind of specifications.
It's also ok to not accept any, but then it should be stated clearly.
Here's an example: https://github.com/ropensci/concstats#development
