### 2023-05-24 NEFSC/READ-PSB-LWT-narwss_rwsas_apps#26 @leahcrowe
> Hi, I did see the line you added. I think it could be more precise for the third party contribution. Often when contributions are accepted there's a code of conduct or some kind of specifications. It's also ok to not accept any, but then it should be stated clearly. Here's an example: https://github.com/ropensci/concstats#development

Thank you for further clarifying these points! I added a "Contributions" section at the end that now includes a Code of Conduct, as well as some further clarifications on submitting issues and code changes.

In working on this, I also added a bit of an intro to the README to better introduce what can be found in the repo, and updated our license so that there is no copyright.
