### 2023-05-28 #5436 @salix-d
Hello,

Regarding the functionality check:
The parts of the app that we have access to do seem to work properly, but a lot of it can't be tested (like all the triggers).

I'm unsure if this is an issue for the paper.
If it is, I did find this package https://rstudio.github.io/shinytest/articles/shinytest.html.
I'm assuming you did do manual testing before submitting, but, maybe having a script showing those tests on your setup might help?
I think it would certainly help in the long run regardless.

And if it's not, then for me, everything on my checklist is checked 😄
