### 2023-05-30 7c53974
Update README.md
don't dupe scripts/
