+++ paper.md
@@ -33,34 +33,35 @@ tags:
 - trigger analysis
 - conservation
 - management
-
 ---
 
 # Summary
 
-Conservation strategies aimed at protecting the Critically Endangered North Atlantic right whale (*Eubalaena glacialis*, hereafter 'right whale') [@CookeIUCN2020] in the United States currently include static and dynamic management measures to mitigate vessel strikes [@NMFS:2013]. This species predominately lives along the eastern seaboard of the United States and Canada where their habitat overlaps with busy shipping ports [@vesselspeedrule2020]. Vessel strikes, along with entanglements in fishing gear, are a leading cause of right whale mortality and the downward trajectory of this species' population size over the last decade [@corkeron2018recovery; @sharp2019gross; @narwc2022]. Seasonal Management Areas (SMAs) are static geographic areas where vessels larger than 65 ft (19.8 m) are required to limit their speed to 10 kts or less at prescribed, annual periods [@NMFS:2008; @NMFS:2013]. Dynamic Right Whale Slow Zones are declared when a threshold density of right whales is detected visually (Dynamic Management Areas, DMAs) [@NMFS:2008; @NMFS:2013] or acoustically (Acoustic Slow Zones) outside of active SMAs [@NOAAFisheries:2020], and within these zones, the vessel speed restriction is voluntary[^1]. The boundaries of dynamic Slow Zones are determined by where the detections occur, and for visual detections, the number of right whales sighted also factors into the calculated size of the area [@NMFS:2008; @NMFS:2013; @NOAAFisheries:2020].  
+Conservation strategies aimed at protecting the Critically Endangered North Atlantic right whale (*Eubalaena glacialis*, hereafter 'right whale') [@CookeIUCN2020] in the United States currently include static and dynamic management measures to mitigate vessel strikes [@NMFS:2013]. This species predominately lives along the eastern seaboard of the United States and Canada where their habitat overlaps with busy shipping ports [@vesselspeedrule2020]. Vessel strikes, along with entanglements in fishing gear, are a leading cause of right whale mortality and the downward trajectory of this species' population size over the last decade [@corkeron2018recovery; @sharp2019gross; @narwc2022]. Seasonal Management Areas (SMAs) are static geographic areas where vessels larger than 65 ft (19.8 m) are required to limit their speed to 10 kts or less at prescribed, annual periods [@NMFS:2008; @NMFS:2013]. Dynamic Right Whale Slow Zones are declared when a threshold density of right whales is detected visually (Dynamic Management Areas, DMAs) [@NMFS:2008; @NMFS:2013] or acoustically (Acoustic Slow Zones) outside of active SMAs [@NOAAFisheries:2020], and within these zones, the vessel speed restriction is voluntary[^1]. The boundaries of dynamic Slow Zones are determined by where the detections occur, and for visual detections, the number of right whales sighted also factors into the calculated size of the area [@NMFS:2008; @NMFS:2013; @NOAAFisheries:2020].
 
[…]
