### 2023-05-31 #5436 @leahcrowe
Thank you for the testing suggestion. I had trouble getting the shinytest to work (I think because of all the html), but I have gotten shinytest2 (https://rstudio.github.io/shinytest2/) to work locally, so I think that will be what we use to demonstrate the functionality specific to being connected to the network server/accessing internal databases.

There are some travel conflicts this week, but our hope is to work this out next week!

Thanks to everyone for you helpful suggestions to improve the project!
