+++ scripts/global_libraries.R
@@ -31,7 +31,8 @@ require(webshot)
 require(rlist)
 require(ellipsis)
 require(stringi)
-require(RODBC)
+#require(RODBC)
+require(ROracle)
 require(DBI)
 require(sqldf)
 require(shinymanager)
