+++ scripts/active_slowzone.R
@@ -39,7 +39,9 @@ if (isolate(criteria$loc) == 'Network') {
   ##action code dataframe to join with results ofvtrigger analysis later
   actioncode <- "select *
             from action"
-  actioncodedf <- sqlQuery(cnxn, actioncode)
+  actioncodedf_q <- dbSendQuery(cnxn, actioncode)
+  actioncodedf<-fetch(actioncodedf_q)
+  
   actioncodedf$ID <- as.numeric(actioncodedf$ID)
   
   #query all relevant DMAs & APZs
@@ -57,7 +59,9 @@ if (isolate(criteria$loc) == 'Network') {
                      and (cancelled not like 'cancel%' or cancelled is null);"
     )
   
-  actdma <- sqlQuery(cnxn, activedmasql)
+  actdma_q <- dbSendQuery(cnxn, activedmasql)
+  actdma<-fetch(actdma_q)
+  
   print(actdma)
   
 } else {
@@ -138,8 +142,8 @@ if (nrow(actdma) == 0) {
                      and (cancelled not like 'cancel%' or cancelled is null);"
     )
     
-    actdma_bounds <- sqlQuery(cnxn, actdma_boundssql)
-    print(actdma_bounds)
+    actdma_bounds_q <- dbSendQuery(cnxn, actdma_boundssql)
+    actdma_bounds<-fetch(actdma_bounds_q)
     
     actdmadf <- actdma %>%
       left_join(actdma_bounds, by = "ID")
