### 2023-06-13 #5436 @leahcrowe
Hello all!

I have added in the dates for manual testing of the Trigger Analysis app in the wiki [here](https://github.com/NEFSC/READ-PSB-LWT-narwss_rwsas_apps/wiki/Trigger-Analysis). For instances where a protection zone is triggered and/or extended, I have also uploaded the reports to those locations.

Unfortunately, I keep running into snags with the shinytest packages. I think it boils down to their requirement of only 1 Rmd file allowed per app. We have several Rmd files in this project, so without substantial workarounds, I don't think I can get this to work (I could get it to work with other projects I have that have only one Rmd).

The main difference between declaring a protection zone through the "Trigger Analysis" route and from sightings data analyzed in the "Aerial Survey Processing App" is the source of the data: "Trigger Analysis" queries data by date from all sources in our Oracle database (both visual and acoustic, opportunistic and dedicated), and the "Aerial Survey Processing App" only looks at sightings from a visual aerial survey. The process that analyzes the detections to see if they meet the criteria for a dynamic protection zone are the same, and occur through the shared files.

Other than record a video of the process, I am not sure what else we can do without substantial code changes. But let me know if you all have other ideas!
