+++ scripts/global_libraries.R
@@ -31,7 +31,8 @@ library(webshot)
 library(rlist)
 library(ellipsis)
 library(stringi)
-library(RODBC)
+#library(RODBC)
 library(DBI)
 library(sqldf)
 library(shinymanager)
+library(ROracle)
