+++ scripts/global_libraries.R
@@ -31,8 +31,7 @@ require(webshot)
 require(rlist)
 require(ellipsis)
 require(stringi)
-#require(RODBC)
-require(ROracle)
+require(RODBC)
 require(DBI)
 require(sqldf)
 require(shinymanager)
