### 2023-07-09 704d827
undo ROracle query
