+++ paper.md
@@ -6,13 +6,14 @@ authors:
   orcid: 0000-0001-9133-8522
   affiliation: 1
 - name: Tim V. N. Cole
+  orcid: 0000-0001-9046-2169
   affiliation: 2
 - name: Heather J. Foley
   orcid: 0000-0003-0898-3008
   affiliation: 1
 - name: Danielle M. Cholewiak
   affiliation: 2
-date: "18 December 2022"
+date: "09 July 2023"
 output:
   word_document: default
   html_document:
