### 2023-07-14 #5436 @graciellehigino
Hi @leahcrowe ! Thank you so much for doing these checks with me. I've skimmed through the manuscript and I only noticed one thing that may be an error: one of the authors in this reference below has only their initials. Can you double-check that too? We're almost there! 🎉

Meyer-Gutbrod, E. L., Greene, C. H., Davies, K. T., & **G, J. D.** (2021). Ocean regime shift is driving collapse of the north atlantic right whale population. Oceanography. https://doi.org/10.5670/oceanog.2021.308
