+++ paper.bib
@@ -126,7 +126,7 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
 }
 
 @article{Meyer-Gutbrod:2021,
-	  author = {Meyer-Gutbrod, Erin L and Greene, Charles H and Davies, Kimberley TA and Johns David G}, 
+	  author = {Meyer-Gutbrod, Erin L and Greene, Charles H and Davies, Kimberley TA and Johns, David G}, 
 	  title = {Ocean Regime Shift is Driving Collapse of the North Atlantic Right Whale Population},
 	  journal = {Oceanography},
 	  year = {2021},
