+++ paper.bib
@@ -8,7 +8,7 @@
   }
  
  @Manual{shiny,
-    title = {shiny: Web Application Framework for R},
+    title = {{shiny: Web Application Framework for R}},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
     year = {2022},
     note = {R package version 1.7.2},
