+++ paper.bib
@@ -76,7 +76,7 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   title={Survey and data considerations to adopt a multi-whale trigger density for protecting the North Atlantic right whale in Canadian waters},
   author={Ratelle, Stephanie M and Hammill, M O and Cole, Tim VN and Hardy, M A and Crowe, Leah M and Elliott, M S},
   journal = {DFO Can. Sci. Advis. Sec. Res. Doc.},
-  year = {in press}}
+  year = {in review}}
 
 @article{baumgartner2019persistent,
   title={Persistent near real-time passive acoustic monitoring for baleen whales from a moored buoy: System description and evaluation},
