+++ paper.bib
@@ -16,7 +16,7 @@
   }
 
 @article{quintana2021residency,
-  title={Residency, demographics, and movement patterns of North Atlantic right whales Eubalaena glacialis in an offshore wind energy development area in southern New England, USA},
+  title={{Residency, demographics, and movement patterns of North Atlantic right whales Eubalaena glacialis in an offshore wind energy development area in southern New England, USA}},
   author={Quintana-Rizzo, E and Leiter, S and Cole, TVN and Hagbloom, MN and Knowlton, AR and Nagelkirk, P and Brien, OO and Khan, CB and Henry, AG and Duley, PA and others},
   journal={Endangered Species Research},
   volume={45},
@@ -26,7 +26,7 @@
 }
 
 @article{DAVIES2019157,
-title = {Mass human-caused mortality spurs federal action to protect endangered North Atlantic right whales in Canada},
+title = {{Mass human-caused mortality spurs federal action to protect endangered North Atlantic right whales in Canada}},
 journal = {Marine Policy},
 volume = {104},
 pages = {157-162},
@@ -73,7 +73,7 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   year={2020}}
 
 @article{ratelle,
-  title={Survey and data considerations to adopt a multi-whale trigger density for protecting the North Atlantic right whale in Canadian waters},
+  title={{Survey and data considerations to adopt a multi-whale trigger density for protecting the North Atlantic right whale in Canadian waters},
   author={Ratelle, Stephanie M and Hammill, M O and Cole, Tim VN and Hardy, M A and Crowe, Leah M and Elliott, M S},
   journal = {DFO Can. Sci. Advis. Sec. Res. Doc.},
   year = {in review}}
@@ -91,7 +91,7 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
 }
 
 @article{parks2011sound,
-  title={Sound production behavior of individual North Atlantic right whales: implications for passive acoustic monitoring},
+  title={{Sound production behavior of individual North Atlantic right whales: implications for passive acoustic monitoring}},
   author={Parks, SE and Searby, A and C{\'e}l{\'e}rier, A and Johnson, MP and Nowacek, DP and Tyack, PL},
   journal={Endangered Species Research},
   volume={15},
@@ -102,7 +102,7 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
 }
 
 @article{corkeron2018recovery,
-  title={The recovery of North Atlantic right whales, Eubalaena glacialis, has been constrained by human-caused mortality},
+  title={{The recovery of North Atlantic right whales, Eubalaena glacialis, has been constrained by human-caused mortality}},
   author={Corkeron, Peter and Hamilton, Philip and Bannister, John and Best, Peter and Charlton, Claire and Groch, Karina R and Findlay, Ken and Rowntree, Victoria and Vermeulen, Els and Pace III, Richard M},
   journal={Royal Society open science},
   volume={5},
@@ -114,7 +114,7 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
 }
 
 @article{davis2017long,
-  title={Long-term passive acoustic recordings track the changing distribution of North Atlantic right whales (Eubalaena glacialis) from 2004 to 2014},
+  title={{Long-term passive acoustic recordings track the changing distribution of North Atlantic right whales (Eubalaena glacialis) from 2004 to 2014}},
[…]
