+++ paper.bib
@@ -56,7 +56,7 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
 }
 
 @article{sharp2019gross,
-  title={Gross and histopathologic diagnoses from North Atlantic right whale Eubalaena glacialis mortalities between 2003 and 2018},
+  title={{Gross and histopathologic diagnoses from North Atlantic right whale Eubalaena glacialis mortalities between 2003 and 2018}},
   author={Sharp, Sarah M and McLellan, William A and Rotstein, David S and Costidis, Alexander M and Barco, Susan G and Durham, Kimberly and Pitchford, Thomas D and Jackson, Katharine A and Daoust, P-Y and Wimmer, Tonya and others},
   journal={Diseases of Aquatic Organisms},
   volume={135},
@@ -73,7 +73,7 @@ author = {Kimberley T.A. Davies and Sean W. Brillant}
   year={2020}}
 
 @article{ratelle,
-  title={{Survey and data considerations to adopt a multi-whale trigger density for protecting the North Atlantic right whale in Canadian waters},
+  title={{Survey and data considerations to adopt a multi-whale trigger density for protecting the North Atlantic right whale in Canadian waters}},
   author={Ratelle, Stephanie M and Hammill, M O and Cole, Tim VN and Hardy, M A and Crowe, Leah M and Elliott, M S},
   journal = {DFO Can. Sci. Advis. Sec. Res. Doc.},
   year = {in review}}
